refactor(types): split agent-hub.ts into runtime and market type modules
