+++ src/lib/types/agent-hub-market.ts
@@ -1,0 +1,15 @@
+export interface AgentMarketCategory {
+  id: string;
+  label: string;
+}
+
+export interface AgentMarketItem {
+  id: string;
+  name: string;
+  summary: string;
+  icon: string;
+  tintColor: string;
+  tags: string[];
+  installsText: string;
+  ratingText: string;
+}
