+++ src/lib/types/agent-hub.ts
@@ -1,3 +1,7 @@
+// Re-export split types for backward compatibility
+export type { RuntimeHostStatus, RuntimeHostRecord, RuntimeServiceStatus } from './agent-hub-runtime';
+export type { AgentMarketCategory, AgentMarketItem } from './agent-hub-market';
+
 // Agent Hub view modes（视图模式）
 export const AGENT_HUB_VIEW_MODES = ['topology', 'list', 'device'] as const;
 export type AgentHubViewMode = (typeof AGENT_HUB_VIEW_MODES)[number];
@@ -80,31 +84,6 @@ export interface AgentDeviceGroup {
   cards: AgentHubDeviceCard[];
 }
 
-// Runtime host status（运行主机状态）用于设备页真实探测展示
-export type RuntimeHostStatus = 'unknown' | 'online' | 'offline' | 'warning';
-
-export interface RuntimeHostRecord {
-  id: string;
-  name: string;
-  host: string;
-  port: number;
-  status: RuntimeHostStatus;
-  createdAt: string;
-  updatedAt: string;
-  lastCheckedAt?: string;
-  lastError?: string;
-  isLocal?: boolean;
-}
-
-export interface RuntimeServiceStatus {
-  running: boolean;
-  host: string;
-  port: number;
-  pid?: number;
-  startedAt?: string;
-  error?: string;
-}
-
 export type AgentAddNodeOptionId = 'input' | 'agent' | 'actor' | 'output' | 'market';
 
 export interface AgentAddNodeOption {
@@ -146,22 +125,6 @@ export interface AgentDetailData {
   triggerRules: AgentDetailKV[];
   targets: AgentHubListItem[];
   recentLogs: AgentDetailTimelineItem[];
-}
-
-export interface AgentMarketCategory {
-  id: string;
-  label: string;
-}
-
-export interface AgentMarketItem {
-  id: string;
-  name: string;
-  summary: string;
-  icon: string;
-  tintColor: string;
-  tags: string[];
-  installsText: string;
-  ratingText: string;
 }
 
 export type AgentConversationRole = 'agent' | 'user';
